chore: clean up v11 code
