+++ packages/react/src/components/Modal/Modal-test.js
@@ -9,7 +9,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import Modal from '../Modal';
+import Modal from './Modal';
 import TextInput from '../TextInput';
 
 describe('Modal', () => {
